# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lumen Server API
  description: API specification and documentation portal for @lumen/server co-signing, fee sponsorship, and policy engine services.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /health:
    get:
      summary: Health Check
      description: Returns the operational status of the server and the configured Stellar network.
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  network:
                    type: string
                    example: testnet
                required:
                  - status
                  - network
  /cosign:
    post:
      summary: Cosign Transaction
      description: Validates transaction operations against active policy rules and appends co-signer signature if approved.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                xdr:
                  type: string
                  description: Base64 encoded Stellar transaction envelope XDR.
                  example: AAAAAG...
                walletAddress:
                  type: string
                  description: Stellar public key address of the originating wallet.
                  example: GAD7654...
              required:
                - xdr
                - walletAddress
      responses:
        '200':
          description: Transaction approved and co-signed
          content:
            application/json:
              schema:
                type: object
                properties:
                  signedXdr:
                    type: string
                    example: AAAAAG...
                required:
                  - signedXdr
        '400':
          description: Invalid request validation error
        '403':
          description: Transaction denied by policy rules
  /fee-bump:
    post:
      summary: Wrap Fee-Bump Transaction
      description: Wraps a transaction envelope in a FeeBumpTransaction signed by the fee sponsor account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                xdr:
                  type: string
                  description: Base64 encoded Stellar transaction envelope XDR to wrap.
                  example: AAAAAG...
              required:
                - xdr
      responses:
        '200':
          description: Fee bump transaction wrapped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  feeBumpXdr:
                    type: string
                    example: AAAAAG...
                required:
                  - feeBumpXdr
  /fee-bump/submit:
    post:
      summary: Submit Fee-Bump Transaction
      description: Wraps and immediately submits a transaction to the Stellar network using the fee sponsor account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                xdr:
                  type: string
                  description: Base64 encoded Stellar transaction envelope XDR.
                  example: AAAAAG...
              required:
                - xdr
      responses:
        '200':
          description: Transaction submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  hash:
                    type: string
                    example: 7f3b...
                  successful:
                    type: boolean
                    example: true
  /wallet/create:
    post:
      summary: Create Seedless Wallet
      description: Provisions a new sponsored Stellar account with multi-signature key configuration.
      responses:
        '200':
          description: Wallet created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                    example: GAD7654...
                  publicKey:
                    type: string
                    example: GAD7654...
                required:
                  - address
                  - publicKey
  /wallet/{address}/analytics:
    get:
      summary: Get Wallet Spending Analytics
      description: Returns aggregated spending analytics for a wallet, sourced from the in-memory policy store trackers. Requires API key authentication.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
          description: Stellar public key address of the wallet.
          example: GAD7654...
      responses:
        '200':
          description: Wallet spending analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletAnalytics'
        '401':
          description: Missing or invalid API key
  /policy:
    post:
      summary: Create or Update Wallet Policy
      description: Configures spend limit, velocity, or allowlist policy rules for a specific wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletId:
                  type: string
                  description: Stellar public key address of the wallet.
                  example: GAD7654...
                rules:
                  type: array
                  items:
                    oneOf:
                      - $ref: '#/components/schemas/SpendLimitRule'
                      - $ref: '#/components/schemas/VelocityRule'
                      - $ref: '#/components/schemas/AllowlistRule'
              required:
                - walletId
                - rules
      responses:
        '200':
          description: Policy created or updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
  /policy/{walletId}:
    get:
      summary: Get Wallet Policy
      description: Retrieves active policy rules for a given wallet address.
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
          description: Stellar public key address of the wallet.
      responses:
        '200':
          description: Wallet policy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '404':
          description: No policy found for wallet
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    WalletAnalytics:
      type: object
      properties:
        address:
          type: string
          example: GAD7654...
        totalTransactions:
          type: integer
          description: Total transactions co-signed for the wallet.
          example: 42
        totalXlmVolume:
          type: string
          description: Total XLM volume co-signed for the wallet.
          example: "1250.0000000"
        policyViolations:
          type: integer
          description: Number of policy violations recorded for the wallet.
          example: 3
        lastTransactionAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the last co-signed transaction, or null if none.
          example: "2024-05-01T12:34:56.000Z"
      required:
        - address
        - totalTransactions
        - totalXlmVolume
        - policyViolations
        - lastTransactionAt
    SpendLimitRule:
      type: object
      properties:
        type:
          type: string
          enum: [spend_limit]
        asset:
          type: string
          example: USDC
        maxPerTx:
          type: string
          example: "100.0"
        maxDaily:
          type: string
          example: "500.0"
      required:
        - type
        - asset
        - maxPerTx
        - maxDaily
    VelocityRule:
      type: object
      properties:
        type:
          type: string
          enum: [velocity]
        maxTransactions:
          type: integer
          example: 10
        windowMinutes:
          type: integer
          example: 60
      required:
        - type
        - maxTransactions
        - windowMinutes
    AllowlistRule:
      type: object
      properties:
        type:
          type: string
          enum: [allowlist]
        destinations:
          type: array
          items:
            type: string
          example: ["GCB2..."]
      required:
        - type
        - destinations
    MaxOperationsRule:
      type: object
      properties:
        type:
          type: string
          enum: [max_operations]
        maxOperations:
          type: integer
          example: 5
      required:
        - type
        - maxOperations
    Policy:
      type: object
      properties:
        id:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        walletId:
          type: string
          example: GAD7654...
        rules:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - walletId
        - rules
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "API key passed in the Authorization: Bearer <api-key> header"
security:
  - bearerAuth: []